refactor: indexation -> indexing
